Function compilers_project_1::parse_regex

pub(crate) fn parse_regex(
    regex: &String, 
    fp_table: &mut HashMap<u32, HashSet<u32>>, 
    s_table: &mut HashMap<char, HashSet<u32>>
) -> Node

loop through input regex to build the Abstract Syntax Tree for the regex. the algorithm used is the one created by Edsger Dijkstra, Shunting Yard Algorithm. https://en.wikipedia.org/wiki/Shunting-yard_algorithm

Input: a regular expression Output: the node that represents the root of the tree.

The followpos table is also updated to be used later in the REGEX -> DFA algorithm.