Function compilers_project_1::regex_dfa

pub(crate) fn regex_dfa(
    fp_table: &HashMap<u32, HashSet<u32>>, 
    s_table: &HashMap<char, HashSet<u32>>, 
    root: &Node, 
    alphabet: &HashSet<char>
) -> Dfa

Constructs a DFA directly from the regular expression.

fp_table: the followpos table that was built while parsing the regular expression. s_table: the symbol table that was built while parsing the regular expression. root: the root Node of the parse tree for the regular expression. alphabet: the alphabet found on the regular expression.

output: DFA.

fp_table and s_table should be built using the parse_regex method. root is also obtained by calling parse_regex.